+++ src/Interpreter.php
@@ -75,7 +75,7 @@ class Interpreter extends \Smuuf\Primi\StrictObject {
 }
 
 ExtensionHub::add([
-	\Smuuf\Primi\Stl\StringExtension::class => \Smuuf\Primi\Structures\StringValue::class,
-	\Smuuf\Primi\Stl\NumberExtension::class => \Smuuf\Primi\Structures\NumberValue::class,
-	\Smuuf\Primi\Stl\ArrayExtension::class => \Smuuf\Primi\Structures\ArrayValue::class,
+	\Smuuf\Primi\Psl\StringExtension::class => \Smuuf\Primi\Structures\StringValue::class,
+	\Smuuf\Primi\Psl\NumberExtension::class => \Smuuf\Primi\Structures\NumberValue::class,
+	\Smuuf\Primi\Psl\ArrayExtension::class => \Smuuf\Primi\Structures\ArrayValue::class,
 ]);
