+++ tests/bootstrap.php
@@ -7,9 +7,9 @@ $loader = new \Smuuf\Koloader\Autoloader(__DIR__ . "/../temp/");
 $loader->addDirectory(__DIR__ . "/../src")->register();
 
 \Smuuf\Primi\ExtensionHub::add([
-	\Smuuf\Primi\Stl\StringExtension::class => \Smuuf\Primi\Structures\StringValue::class,
-	\Smuuf\Primi\Stl\NumberExtension::class => \Smuuf\Primi\Structures\NumberValue::class,
-	\Smuuf\Primi\Stl\ArrayExtension::class => \Smuuf\Primi\Structures\ArrayValue::class,
+	\Smuuf\Primi\Psl\StringExtension::class => \Smuuf\Primi\Structures\StringValue::class,
+	\Smuuf\Primi\Psl\NumberExtension::class => \Smuuf\Primi\Structures\NumberValue::class,
+	\Smuuf\Primi\Psl\ArrayExtension::class => \Smuuf\Primi\Structures\ArrayValue::class,
 ]);
 
 \Tester\Environment::setup();
